backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Werkzeug==2.0.1
redis==4.0.2
celery
Jinja2==3.0.1
SQLAlchemy==1.4.23
email-validator==1.1.3
python-dotenv==0.19.0
setuptools
flask_mail
psycopg2-binary==2.9.9
Flask-Migrate==4.0.5